+++ src/base64.rs
@@ -12,6 +12,78 @@ use crate::Error;
 const ALPHABET_BASE64URL: &str = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_";
 const ALPHABET_BASE64URL_BYTES: &[u8] = ALPHABET_BASE64URL.as_bytes();
 
+fn decode_char(c: char) -> Result<u8, Error> {
+    let idx = match c {
+        'A' => 0,
+        'B' => 1,
+        'C' => 2,
+        'D' => 3,
+        'E' => 4,
+        'F' => 5,
+        'G' => 6,
+        'H' => 7,
+        'I' => 8,
+        'J' => 9,
+        'K' => 10,
+        'L' => 11,
+        'M' => 12,
+        'N' => 13,
+        'O' => 14,
+        'P' => 15,
+        'Q' => 16,
+        'R' => 17,
+        'S' => 18,
+        'T' => 19,
+        'U' => 20,
+        'V' => 21,
+        'W' => 22,
+        'X' => 23,
+        'Y' => 24,
+        'Z' => 25,
+        'a' => 26,
+        'b' => 27,
+        'c' => 28,
+        'd' => 29,
+        'e' => 30,
+        'f' => 31,
+        'g' => 32,
+        'h' => 33,
+        'i' => 34,
+        'j' => 35,
+        'k' => 36,
+        'l' => 37,
+        'm' => 38,
+        'n' => 39,
+        'o' => 40,
+        'p' => 41,
+        'q' => 42,
+        'r' => 43,
+        's' => 44,
+        't' => 45,
+        'u' => 46,
+        'v' => 47,
+        'w' => 48,
+        'x' => 49,
+        'y' => 50,
+        'z' => 51,
+        '0' => 52,
+        '1' => 53,
+        '2' => 54,
+        '3' => 55,
+        '4' => 56,
+        '5' => 57,
+        '6' => 58,
+        '7' => 59,
+        '8' => 60,
+        '9' => 61,
+        '-' => 62,
+        '_' => 63,
+        _ => return Err(Error::InvalidCharacter),
+    };
+
+    Ok(idx)
+}
+
 #[rustfmt::skip]
 pub fn encode_i64(input: i64) -> [char; 11] {
     let b = input.to_be_bytes();
@@ -58,11 +130,7 @@ pub fn decode_i64(input: [char; 11]) -> Result<i64, Error> {
     let mut c: [u8; 11] = [0; 11];
 
     for i in 0..=10 {
-        let idx = ALPHABET_BASE64URL
-            .find(input[i])
-            .ok_or(Error::InvalidCharacter)?;
-
-        c[i] = u8::try_from(idx).map_err(Error::InfallibleU8FromUsize)?;
+        c[i] = decode_char(input[i])?;
     }
 
     let p1 = decode_quantum([c[0], c[1], c[2], c[3]]);
@@ -81,11 +149,7 @@ pub fn decode_i32(input: [char; 6]) -> Result<i32, Error> {
     let mut c: [u8; 6] = [0; 6];
 
     for i in 0..=5 {
-        let idx = ALPHABET_BASE64URL
-            .find(input[i])
-            .ok_or(Error::InvalidCharacter)?;
-        
-        c[i] = u8::try_from(idx).map_err(Error::InfallibleU8FromUsize)?;
+        c[i] = decode_char(input[i])?;
     }
 
     let p1 = decode_quantum([c[0], c[1], c[2], c[3]]);
@@ -102,11 +166,7 @@ pub fn decode_i16(input: [char; 3]) -> Result<i16, Error> {
     let mut c: [u8; 3] = [0; 3];
 
     for i in 0..=2 {
-        let idx = ALPHABET_BASE64URL
-            .find(input[i])
-            .ok_or(Error::InvalidCharacter)?;
-
-        c[i] = u8::try_from(idx).map_err(Error::InfallibleU8FromUsize)?;
+        c[i] = decode_char(input[i])?;
     }
 
     let p1 = decode_partial_16(c)?;