+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Id32` struct for 32 bit value support
 - `Id16` struct for 16 bit value support
 
+### Removed
+- (Breaking) Removed InfallibleU8FromUsize Error variant
+
 ### Changes
 - Updated explanatory text and code examples in README.md
 - Updated rustdoc documentation